feat: add change structure
