+++ .gitignore
@@ -1,1 +1,3 @@
-node_modules
+**/.DS_Store
+/node_modules
+/dist
