+++ config/webpack.common.js
@@ -1,0 +1,76 @@
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+// const CopyWebpackPlugin = require('copy-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+const paths = require('./paths');
+
+module.exports = {
+  // Where webpack looks to start building the bundle
+  entry: './src/index.js',
+
+  // Where webpack outputs the assets and bundles
+  output: {
+    path: paths.build,
+    filename: '[name].bundle.js',
+    publicPath: '',
+    assetModuleFilename: 'assets/images/[name][ext]',
+  },
+
+  // Customize the webpack build process
+  plugins: [
+    // Removes/cleans build folders and unused assets when rebuilding
+    new CleanWebpackPlugin(),
+
+    // Copies files from target to destination folder
+    // new CopyWebpackPlugin({
+    //   patterns: [
+    //     {
+    //       from: `${paths.src}/assets/images/`,
+    //       to: `${paths.build}/assets/images`,
+    //       globOptions: {
+    //         ignore: ['*.DS_Store'],
+    //       },
+    //     },
+    //   ],
+    // }),
+
+    // Generates an HTML file from a template
+    // Generates deprecation warning: https://github.com/jantimon/html-webpack-plugin/issues/1501
+    new HtmlWebpackPlugin({
+      // favicon: paths.src + '/images/favicon.png',
+      template: `${paths.src}/index.html`, // template file
+      filename: 'index.html', // output file
+    }),
+  ],
+
+  // Determine how modules within the project are treated
+  module: {
+    rules: [
+      // JavaScript: Use Babel to transpile JavaScript files
+      { test: /\.js$/, exclude: /node_modules/, use: ['babel-loader'] },
+
+      // Styles: Inject CSS into the head with source maps
+      {
+        test: /\.(scss|css)$/,
+        use: [
+          'style-loader',
+          { loader: 'css-loader', options: { sourceMap: true, importLoaders: 1 } },
+          { loader: 'postcss-loader', options: { sourceMap: true } },
+          { loader: 'sass-loader', options: { sourceMap: true } },
+        ],
+      },
+
+      // Images: Copy image files to build folder
+      {
+        test: /\.(?:ico|gif|png|jpg|jpeg)$/i,
+        type: 'asset/resource',
+        // generator: {
+        //   filename: 'assets/[name][ext]',
+        // },
+      },
+
+      // Fonts and SVGs: Inline files
+      { test: /\.(woff(2)?|eot|ttf|otf|svg|)$/, type: 'asset/inline' },
+    ],
+  },
+};
