+++ config/webpack.dev.js
@@ -1,0 +1,27 @@
+const webpack = require('webpack');
+const { merge } = require('webpack-merge');
+const paths = require('./paths');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  // Set the mode to development or production
+  mode: 'development',
+
+  // Control how source maps are generated
+  devtool: 'inline-source-map',
+
+  // Spin up a server for quick development
+  devServer: {
+    historyApiFallback: true,
+    contentBase: paths.src,
+    open: true,
+    compress: true,
+    // hot: true,
+    port: 8080,
+  },
+
+  plugins: [
+    // Only update what has changed on hot reload
+    new webpack.HotModuleReplacementPlugin(),
+  ],
+});
