+++ package.json
@@ -10,15 +10,36 @@
     "build": "cross-env NODE_ENV=production webpack --config config/webpack.prod.js"
   },
   "devDependencies": {
-    "css-loader": "^5.0.1",
-    "eslint": "^7.15.0",
-    "eslint-config-airbnb-base": "^14.2.1",
+    "@babel/core": "^7.12.1",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@types/codemirror": "0.0.98",
+    "babel-loader": "^8.1.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^6.2.1",
+    "cross-env": "^7.0.2",
+    "css-loader": "^5.0.0",
+    "css-minimizer-webpack-plugin": "^1.1.5",
+    "eslint": "^7.12.1",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-html": "^6.1.0",
     "eslint-plugin-import": "^2.22.1",
-    "node-sass": "^5.0.0",
-    "sass-loader": "^10.1.0",
+    "html-webpack-plugin": "^5.0.0-alpha.7",
+    "mini-css-extract-plugin": "^1.0.0",
+    "node-sass": "^4.14.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.4",
+    "postcss-loader": "^4.0.4",
+    "postcss-preset-env": "^6.7.0",
+    "sass-loader": "^10.0.3",
     "style-loader": "^2.0.0",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.5.0",
-    "webpack-cli": "^4.2.0"
+    "webpack": "^5.1.3",
+    "webpack-cli": "^4.0.0",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-merge": "^5.2.0"
+  },
+  "dependencies": {
+    "leaflet": "^1.7.1",
+    "regenerator-runtime": "^0.13.7"
   }
 }
