+++ postcss.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: {
+    'postcss-preset-env': {
+      browsers: 'last 2 versions',
+    },
+  },
+}
