+++ src/scripts/map/Model/index.js
