+++ src/scripts/map/View/index.js
