+++ src/scripts/map/constants/index.js
