+++ src/scripts/state.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-await-in-loop */
+import regeneratorRuntime from 'regenerator-runtime';
 // eslint-disable-next-line import/no-mutable-exports
 export let covidData;
 // eslint-disable-next-line import/no-mutable-exports
@@ -10,8 +12,8 @@ export const StateClass = class {
             this.data = await this.res.json();
             if (this.data.Message === '') {
                 covidData = this.data;
-                for (this.i = 0; i < covidData.Countries.length; this.i += 1) {
-                    this.urlMapCenter = `https://api.mapbox.com/geocoding/v5/mapbox.places/${dataAll.Countries[i].Country},${dataAll.Countries[i].CountryCode}.json?types=country&access_token=pk.eyJ1IjoibXlmZW5peDkyIiwiYSI6ImNrYXBpdXhwMTF5NTYzMXA2emY0M3pnd24ifQ.I73eBezMUvPr3OAN-aF1Cg`;
+                for (this.i = 0; this.i < covidData.Countries.length; this.i += 1) {
+                    this.urlMapCenter = `https://api.mapbox.com/geocoding/v5/mapbox.places/${covidData.Countries[this.i].Country},${covidData.Countries[this.i].CountryCode}.json?types=country&access_token=pk.eyJ1IjoibXlmZW5peDkyIiwiYSI6ImNrYXBpdXhwMTF5NTYzMXA2emY0M3pnd24ifQ.I73eBezMUvPr3OAN-aF1Cg`;
                     this.resMapCenter = await fetch(this.urlMapCenter);
                     this.dataMapCenter = await this.resMapCenter.json();
                     covidData.Countries[this.i].centerCountry = this.dataMapCenter.features[0].center.reverse();
